Move BitStream to //types
