+++ fly/coders/coder.cpp
@@ -1,7 +1,7 @@
 #include "fly/coders/coder.h"
 
-#include "fly/coders/bit_stream.h"
 #include "fly/logger/logger.h"
+#include "fly/types/bit_stream/bit_stream.h"
 
 #include <chrono>
 #include <fstream>
