+++ test/coders/huffman.cpp
@@ -1,9 +1,9 @@
-#include "fly/coders/bit_stream.h"
 #include "fly/coders/huffman/huffman_config.h"
 #include "fly/coders/huffman/huffman_decoder.h"
 #include "fly/coders/huffman/huffman_encoder.h"
 #include "fly/literals.h"
-#include "fly/types/string.h"
+#include "fly/types/bit_stream/bit_stream.h"
+#include "fly/types/string/string.h"
 #include "test/util/path_util.h"
 
 #include <gtest/gtest.h>
